refactor: extract common code from wallets
