+++ src/base/safewallet.ts
@@ -1,13 +1,10 @@
 import {
-  SafeTransactionDataPartial,
-  SafeMultisigTransactionResponse,
-  SafeMultisigConfirmationResponse,
   MetaTransactionData,
 } from "@safe-global/safe-core-sdk-types";
 import Safe, { EthersAdapter } from "@safe-global/protocol-kit";
 import SafeApiKit from "@safe-global/api-kit";
 import { ethers, Wallet, HDNodeWallet } from "ethers";
-const ApiKit = require("@safe-global/api-kit");
+import { concatSignatures, isTransactionSignedByAddress } from "./wallet";
 
 type Opts = {
   allowedDomains?: RegExp[];
@@ -36,6 +33,7 @@ export class SafeWallet {
   public signer: Wallet | HDNodeWallet;
   private safeSdk: Safe;
   private safeService: SafeApiKit;
+
   constructor(
     address: string,
     apiService: string,
@@ -62,40 +60,6 @@ export class SafeWallet {
     });
   }
 
-  private isTransactionSignedByAddress(
-    tx: SafeMultisigTransactionResponse
-  ): boolean {
-    const confirmation = tx.confirmations.find(
-      (confirmation) => confirmation.owner === this.signer.address
-    );
-    return !!confirmation;
-  }
-
-  private concatSignatures(tx: SafeMultisigTransactionResponse): string | null {
-    if (tx.confirmations.length < tx.confirmationsRequired) {
-      return null;
-    }
-    // must sort by address
-    tx.confirmations.sort(
-      (
-        left: SafeMultisigConfirmationResponse,
-        right: SafeMultisigConfirmationResponse
-      ) => {
-        const leftAddress = left.owner.toUpperCase();
-        const rightAddress = right.owner.toUpperCase();
-        if (leftAddress < rightAddress) {
-          return -1;
-        } else {
-          return 1;
-        }
-      }
-    );
-    var signatures = "0x";
-    for (const confirmation of tx.confirmations) {
-      signatures += confirmation.signature.substring(2);
-    }
-    return signatures;
-  }
 
   async proposeTransaction(
     transactions: MetaTransactionData[],
@@ -107,8 +71,8 @@ export class SafeWallet {
     const safeTxHash = await this.safeSdk.getTransactionHash(tx);
     try {
       const transaction = await this.safeService.getTransaction(safeTxHash);
-      var signatures = this.concatSignatures(transaction);
-      const hasBeenSigned = this.isTransactionSignedByAddress(transaction);
+      var signatures = concatSignatures(transaction);
+      const hasBeenSigned = isTransactionSignedByAddress(transaction);
       if (hasBeenSigned || signatures !== null) {
         //const isValidTx = await this.safeSdk.isValidTransaction(transaction);
         return {
